refactor: move ProgramCache to its own file
