+++ systems/renderer/basic.js
@@ -1,7 +1,7 @@
 import { pipeline as SHADERS } from "pex-shaders";
 
 import createBaseSystem from "./base.js";
-import { ProgramCache } from "../../utils.js";
+import ProgramCache from "../../program-cache.js";
 
 // Impacts program caching
 // prettier-ignore
