+++ systems/renderer/helper.js
@@ -3,7 +3,8 @@ import { pipeline as SHADERS } from "pex-shaders";
 import createGeomBuilder from "geom-builder";
 
 import createBaseSystem from "./base.js";
-import { ProgramCache, TEMP_MAT4, TEMP_VEC3 } from "../../utils.js";
+import { TEMP_MAT4, TEMP_VEC3 } from "../../utils.js";
+import ProgramCache from "../../program-cache.js";
 
 // Impacts program caching
 // prettier-ignore
