+++ systems/renderer/standard.js
@@ -2,7 +2,9 @@ import { mat3, mat4 } from "pex-math";
 import { pipeline as SHADERS } from "pex-shaders";
 
 import createBaseSystem from "./base.js";
-import { NAMESPACE, ProgramCache, TEMP_MAT4 } from "../../utils.js";
+import { NAMESPACE, TEMP_MAT4 } from "../../utils.js";
+import ProgramCache from "../../program-cache.js";
+
 import * as AreaLightsData from "./area-light-data.js";
 
 let ltc_1;
